# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de Compte",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes",
                "operationId": "03ebbdde72897aa34cb2d14fa567a91e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre par page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cheque",
                                "epargne"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tri",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "numero_compte"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comptes/archived": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes archivés",
                "operationId": "78c8032756b6560bad0a4841d6240166",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre par page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cheque",
                                "epargne"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes archivés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comptes/{id}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Détails d'un compte",
                "operationId": "73c661e1b184667f1bb194f22d82cb7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Compte": {
                "properties": {
                    "id": {
                        "description": "ID du compte",
                        "type": "string"
                    },
                    "numeroCompte": {
                        "description": "Numéro du compte",
                        "type": "string"
                    },
                    "titulaire": {
                        "description": "Nom du titulaire",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "cheque",
                            "epargne"
                        ]
                    },
                    "solde": {
                        "description": "Solde calculé",
                        "type": "number",
                        "format": "float"
                    },
                    "devise": {
                        "description": "Devise",
                        "type": "string"
                    },
                    "dateCreation": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date"
                    },
                    "statut": {
                        "description": "Statut",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ]
                    },
                    "metadata": {
                        "$ref": "#/components/schemas/Metadata"
                    }
                },
                "type": "object"
            },
            "Metadata": {
                "properties": {
                    "derniereModification": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "version": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "currentPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "itemsPerPage": {
                        "type": "integer"
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Links": {
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Gestion des comptes bancaires"
        }
    ]
}